perf(ptv3): use uint32 serialized metadata

+++ perception/autoware_ptv3/include/autoware/ptv3/preprocess/preprocess_kernel.hpp
@@ -30,14 +30,14 @@ namespace autoware::ptv3
 
 struct SerializedPoolingDeviceStageView
 {
-  std::int64_t * indices{};
-  std::int64_t * indptr{};
-  std::int64_t * head_indices{};
-  std::int64_t * cluster{};
+  std::uint32_t * indices{};
+  std::uint32_t * indptr{};
+  std::uint32_t * head_indices{};
+  std::uint32_t * cluster{};
   std::int32_t * grid_coord{};
-  std::int64_t * serialized_code{};
-  std::int64_t * serialized_order{};
-  std::int64_t * serialized_inverse{};
+  std::uint32_t * serialized_code{};
+  std::uint32_t * serialized_order{};
+  std::uint32_t * serialized_inverse{};
 };
 
 class PreprocessCuda
@@ -47,13 +47,14 @@ public:
 
   std::size_t generateFeatures(
     const void * input_data, CloudFormat input_format, unsigned int num_points,
-    float * voxel_features, std::int32_t * voxel_coords, std::int64_t * voxel_hashes,
+    float * voxel_features, std::int32_t * voxel_coords, std::uint32_t * voxel_hashes,
     void * compact_points, float * reconstruction_features, void * cropped_source_points,
     std::int64_t * inverse_map, std::size_t * num_cropped_points);
 
   void generateSerializedPoolingMetadata(
-    const std::int32_t * grid_coord, const std::int64_t * serialized_code, std::int64_t num_voxels,
-    const std::vector<SerializedPoolingDeviceStageView> & stages, std::int64_t * stage_counts);
+    const std::int32_t * grid_coord, const std::uint32_t * serialized_code,
+    std::uint32_t num_voxels, const std::vector<SerializedPoolingDeviceStageView> & stages,
+    std::uint32_t * stage_counts);
 
   [[nodiscard]] const std::uint32_t * cropMask() const { return crop_mask_d_.get(); }
   [[nodiscard]] const std::uint32_t * cropIndices() const { return crop_indices_d_.get(); }
@@ -85,12 +86,12 @@ private:
   autoware::cuda_utils::CudaUniquePtr<std::uint8_t[]> sort_workspace_d_{nullptr};
   std::size_t sort_workspace_size_{0};
 
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_keys_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_sorted_keys_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_indices_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_sorted_indices_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_run_flags_d_{nullptr};
-  autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_run_ids_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_keys_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_sorted_keys_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_indices_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_sorted_indices_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_run_flags_d_{nullptr};
+  autoware::cuda_utils::CudaUniquePtr<std::uint32_t[]> pooling_run_ids_d_{nullptr};
   autoware::cuda_utils::CudaUniquePtr<std::uint8_t[]> pooling_workspace_d_{nullptr};
   std::size_t pooling_workspace_size_{0};
 };
